mobile/android/tests/browser/chrome/test_referrer.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1549732
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1549732</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript">

  "use strict";

  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");
  ChromeUtils.import("resource://gre/modules/Messaging.jsm");

  // The chrome window
  let chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let BrowserApp = chromeWin.BrowserApp;

  // Track the tabs where the tests are happening
  let tabTest;

  function cleanupTabs() {
    if (tabTest) {
      BrowserApp.closeTab(tabTest);
      tabTest = null;
    }
  }

  SimpleTest.registerCleanupFunction(function() {
    cleanupTabs();
  });

  const kTestPage = "https://example.com";
  const kReferrer = "https://foo.org/";

  add_task(async function test_referrer() {
    tabTest = BrowserApp.addTab(kTestPage, { referrerURI: kReferrer,
                                              parentId: BrowserApp.selectedTab.id,
                                              selected: true});
    await promiseBrowserEvent(tabTest.browser, "DOMContentLoaded");

    // Check that basic_article is now selected
    is(BrowserApp.selectedBrowser, tabTest.browser, "Target tab is selected after being added.");
    is(tabTest.browser.contentDocument.referrer, kReferrer, "Target tab has correct referrer");

    cleanupTabs();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1549732">Mozilla Bug 1549732</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
